fix: Move import to top of file (ruff E402)
