+++ src/azlin/remote_exec.py
@@ -14,8 +14,9 @@ import logging
 import subprocess
 from concurrent.futures import ThreadPoolExecutor, as_completed
 from dataclasses import dataclass
-from typing import Any
 
+# TmuxSession moved to azlin.models to avoid circular dependency
+from azlin.models.session_models import TmuxSession
 from azlin.modules.ssh_connector import SSHConfig
 from azlin.modules.ssh_routing_resolver import SSHRoute
 
@@ -265,38 +266,6 @@ class RemoteExecutor:
             slug = slug[:max_length].rstrip("-")
 
         return slug if slug else "cmd"
-
-
-@dataclass
-class TmuxSession:
-    """Information about a tmux session."""
-
-    vm_name: str
-    session_name: str
-    windows: int
-    created_time: str
-    attached: bool = False
-
-    def to_dict(self) -> dict[str, Any]:
-        """Convert to dictionary for caching."""
-        return {
-            "vm_name": self.vm_name,
-            "session_name": self.session_name,
-            "windows": self.windows,
-            "created_time": self.created_time,
-            "attached": self.attached,
-        }
-
-    @classmethod
-    def from_dict(cls, data: dict[str, Any]) -> "TmuxSession":
-        """Create from cached dictionary."""
-        return cls(
-            vm_name=data["vm_name"],
-            session_name=data["session_name"],
-            windows=data["windows"],
-            created_time=data["created_time"],
-            attached=data.get("attached", False),
-        )
 
 
 class TmuxSessionExecutor:
